refactor(init): move resolveTypeclawSpec into cli-version
Relocate `resolveTypeclawSpec` (and a small `typeclawCheckoutRoot` helper) from
index.ts into cli-version.ts, where the CLI-vs-checkout detection already lives.
This lets the start-time auto-reconcile (next commit) reuse the same spec
resolver without importing the heavy init barrel (which would cycle through
@/container). No behavior change: the spec it returns is identical — `^version`
for an installed CLI, `link:typeclaw` on native-Windows dev, `file:<rel>` on
POSIX dev — and init scaffolding + the Windows bun-link step call through it.
